museum-backend/requirements.txt
fastapi
uvicorn
psycopg2
python-dotenv
bcrypt==4.0.1
pydantic
email-validator

